lwip Library Port
Added the lwip library as a standalone module (i.e. without linking to
the Stellaris folder). Updated the dependencies of the project to
correctly point to these files.

Increased the stack size from 512 to 1024 because the board was running
out of memory. (Checked the ethernet example and they also use a stack
of 1024)

lm3s8962_startup_ccs.c: Linked interrupt handlers for the Systick and
the ethernet module.

Main: Added the following methods:
DisplayIPAddress
SysTickIntHandler
lwIPHostTimerHandler
ftp_init/ftp_accept: These methods will probably move to other file, but
they were used as a quick test to see if everything was working after
porting the library.