Implemented Part of Protocol Interpreter
Main: Removed the hacky methods that were  used to test the TCP API.

FtpProtocolInterpreter.c:
Added ftp_SendMsg: to handle the transmission of FTP responses.
Added ftp_RxCmd: this method handles the reception of packages.
Added ftp_CmdSent: this method is called when the TCP module transmits a
packet.
Added ftp_Poll: this method gets called when the TCP timers expire. It
is now empty, but it should be used to check the state of the data
transfer requests.
Added ftp_PiError: this function gets called when an error occurs. Right
now it is empty.
Added ftp_Accept: this method completes the connection between the TCP
module and the FTP module. This is called when TCP receives an FTP
request.
Added ftp_Init: this method is used to initialize the TCP module to
listen for FTP requests.

FtpProtocolInterpreter.h: Added a short list of FTP messages and the
methods that need to be available for other modules.