Data Connection Fixed
I have fixed the issue with the data connection where it was printing
garbage from the strings.  I fixed this using the parsePrintableString
as suggested by Clayton.

The STOR command now completes successfully. It does not actually write
a file, but it prints its contents through the debug UART.

The parsePrintableString method adds a null terminator to mark the end
of the parsed string.